Adding detail to the data model
