+++ experiments/ambleck/src/main/java/uk/co/metagrid/ambleck/datamodel/resource/compute/ComputeResource.java
@@ -22,6 +22,7 @@
  */
 package uk.co.metagrid.ambleck.datamodel.resource.compute;
 
+import uk.co.metagrid.ambleck.datamodel.AbstractObject;
 import uk.co.metagrid.ambleck.datamodel.resource.AbstractResource;
 
 public class ComputeResource extends AbstractResource {
@@ -36,6 +37,80 @@ public class ComputeResource extends AbstractResource {
     public ComputeResource(final String name)
         {
         super(name, TYPE_URL);
+        }
+
+    public static class MinMaxPair<Type> {
+
+        public MinMaxPair()
+            {
+            }
+
+        public MinMaxPair(final Type min, final Type max)
+            {
+            this.min = min;
+            this.max = max;
+            }
+
+        private Type min ;
+        public Type getMin()
+            {
+            return this.min ;
+            }
+        public void setMin(final Type min)
+            {
+            this.min = min;
+            }
+        private Type max ;
+        public Type getMax()
+            {
+            return this.max ;
+            }
+        public void setMax(final Type max)
+            {
+            this.max = max;
+            }
+        }
+
+    public static class MinMaxInteger extends MinMaxPair<Integer> {
+
+        }
+
+    public static class MinMaxFloat extends MinMaxPair<Float> {
+
+        }
+
+    public static class ComputeSpecific extends AbstractObject.AbstractSpecific {
+
+        public ComputeSpecific()
+            {
+            super();
+            }
+
+        public ComputeSpecific(final MinMaxFloat cores)
+            {
+            super();
+            this.cores = cores ;
+            }
+
+        private MinMaxFloat cores;
+        public MinMaxFloat getCores()
+            {
+            return this.cores ;
+            }
+        public void setCores(final MinMaxFloat cores)
+            {
+            this.cores = cores ;
+            }
+        }
+
+    private ComputeSpecific spec ;
+    public ComputeSpecific getSpec()
+        {
+        return this.spec;
+        }
+    public void setSpec(final ComputeSpecific spec)
+        {
+        this.spec = spec;
         }
