+++ notes/zrq/20240402-02-planning.txt
@@ -1,0 +1,46 @@
+#
+# <meta:header>
+#   <meta:licence>
+#     Copyright (c) 2024, Manchester (http://www.manchester.ac.uk/)
+#
+#     This information is free software: you can redistribute it and/or modify
+#     it under the terms of the GNU General Public License as published by
+#     the Free Software Foundation, either version 3 of the License, or
+#     (at your option) any later version.
+#
+#     This information is distributed in the hope that it will be useful,
+#     but WITHOUT ANY WARRANTY; without even the implied warranty of
+#     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#     GNU General Public License for more details.
+#
+#     You should have received a copy of the GNU General Public License
+#     along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#   </meta:licence>
+# </meta:header>
+#
+#zrq-notes-indent
+#
+# AIMetrics: []
+#
+
+    Target:
+
+        Team teal-c planning meeting. 
+
+    Result:
+
+        Work in progress ...
+
+# -----------------------------------------------------
+
+
+Meeting Zoom Tuesday 13:30
+https://skatelescope.zoom.us/j/5104793729?pwd=a093cFlqaVJFdFBMblg5SmNiY0Mvdz09
+
+Miro board
+https://miro.com/app/board/uXjVNnnrzyE
+
+Kanban board
+https://jira.skatelescope.org/secure/RapidBoard.jspa?rapidView=494&quickFilter=1788
+
+
