+++ notes/zrq/20240402-03-libgtk3-nocsd.txt
@@ -1,0 +1,78 @@
+#
+# <meta:header>
+#   <meta:licence>
+#     Copyright (c) 2024, Manchester (http://www.manchester.ac.uk/)
+#
+#     This information is free software: you can redistribute it and/or modify
+#     it under the terms of the GNU General Public License as published by
+#     the Free Software Foundation, either version 3 of the License, or
+#     (at your option) any later version.
+#
+#     This information is distributed in the hope that it will be useful,
+#     but WITHOUT ANY WARRANTY; without even the implied warranty of
+#     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#     GNU General Public License for more details.
+#
+#     You should have received a copy of the GNU General Public License
+#     along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#   </meta:licence>
+# </meta:header>
+#
+#zrq-notes-indent
+#
+# AIMetrics: []
+#
+
+    Target:
+
+        Fix issues with laptop configuration.
+
+    Result:
+
+        Work in progress ...
+
+# -----------------------------------------------------
+# History ignores commands starting with white space.
+#[user@laptop]
+
+
+    Fix - set HISTCONTROL in ~/.bashrc
+
+    https://unix.stackexchange.com/a/483705
+
+# -----------------------------------------------------
+# Running yq reports an error.
+#[user@laptop]
+
+    yq ...
+
+    >   ERROR: ld.so: object 'libgtk3-nocsd.so.0' from LD_PRELOAD cannot be preloaded (failed to map segment from shared object): ignored.
+    >   ....
+    >   ....
+
+    # Temp fix is to unset the environment variable.
+    # Not long term, because we want gtk3-nocsd to run.
+    echo $LD_PRELOAD
+
+    >   libgtk3-nocsd.so.0
+
+    unset LD_PRELOAD
+
+    yq ...
+
+    >   ....
+    >   ....
+
+        Resources :
+
+            https://github.com/PCMan/gtk3-nocsd
+
+            https://askubuntu.com/questions/1462779/object-libgtk3-nocsd-so-0-from-ld-preload-cannot-be-preloaded-when-running-chr
+
+
+        Note - yq was installed via snap, so it might be a library version issue.
+
+
+
+
+
